requirements.txt
mysql-connector-python==8.0.32
pika==1.3.1
pika-stubs==0.1.3
python-dotenv==1.0.0
watchdog==2.3.1
psycopg2==2.9.3
pytest==7.2.2